requirements.txt
# Web Scraping
beautifulsoup4==4.12.2
scrapy==2.11.0
selenium==4.15.2
requests==2.31.0
lxml==4.9.3

# Data Processing
pandas==2.1.3
numpy==1.26.2

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Visualization
streamlit==1.28.2
plotly==5.18.0
matplotlib==3.8.2
seaborn==0.13.0

# Utilities
python-dotenv==1.0.0
schedule==1.2.0
loguru==0.7.2

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Development
black==23.11.0
flake8==6.1.0
